Refactored revocation to revocation manager and all receipts methods to receipt repository along with some minor refactorings and changes
