+++ sdk/src/main/java/com/microsoft/did/sdk/RevocationManager.kt
@@ -1,47 +1,74 @@
-// Copyright (c) Microsoft Corporation. All rights reserved
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
 
 package com.microsoft.did.sdk
 
 import com.microsoft.did.sdk.credential.models.VerifiableCredentialHolder
 import com.microsoft.did.sdk.credential.models.receipts.ReceiptAction
-import com.microsoft.did.sdk.credential.service.models.RevokedRPNameAndDid
+import com.microsoft.did.sdk.credential.service.models.RpDidToNameMap
+import com.microsoft.did.sdk.datasource.repository.ReceiptRepository
 import com.microsoft.did.sdk.datasource.repository.VerifiableCredentialHolderRepository
 import com.microsoft.did.sdk.util.controlflow.Result
 import com.microsoft.did.sdk.util.controlflow.runResultTry
-import com.microsoft.did.sdk.util.createAndSaveReceiptsForVCs
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import javax.inject.Inject
+import javax.inject.Singleton
 
-class RevocationManager (private val vchRepository: VerifiableCredentialHolderRepository) {
+/**
+ * This class manages revocation of Verifiable Presentation(s)
+ */
+@Singleton
+class RevocationManager @Inject constructor(
+    private val vchRepository: VerifiableCredentialHolderRepository,
+    private val receiptRepository: ReceiptRepository
+) {
 
     /**
-     * Revokes a verifiable presentation which revokes access for specific relying party/parties or all relying parties to do a status check on the Verifiable Credential
+     * Revokes verifiable presentation(s) which revokes access for specific relying party/parties to do a status check on the Verifiable Credential
      *
      * @param verifiableCredentialHolder The VC for which access to check status is revoked
-     * @param rpDidAndName Map of DIDs and names of relying parties whose access is revoked
+     * @param rpDidToNameMap Map of DIDs and names of relying parties whose access is revoked
      * @param reason Reason for revocation
      */
-
     suspend fun revokeVerifiablePresentation(
         verifiableCredentialHolder: VerifiableCredentialHolder,
-        rpDidAndName: RevokedRPNameAndDid,
+        rpDidToNameMap: RpDidToNameMap,
         reason: String = ""
     ): Result<Unit> {
         return withContext(Dispatchers.IO) {
             runResultTry {
-                vchRepository.revokeVerifiablePresentation(verifiableCredentialHolder, rpDidAndName.keys.toList(), reason).abortOnError()
-                if (rpDidAndName.isEmpty())
-                    createAndSaveReceiptsForVCs("", "", ReceiptAction.Revocation, listOf(verifiableCredentialHolder.cardId), vchRepository)
-                else
-                    rpDidAndName.forEach { relyingParty ->
-                        createAndSaveReceiptsForVCs(
-                            relyingParty.key,
-                            relyingParty.value,
-                            ReceiptAction.Revocation,
-                            listOf(verifiableCredentialHolder.cardId),
-                            vchRepository
-                        )
-                    }
+                vchRepository.revokeVerifiablePresentation(verifiableCredentialHolder, rpDidToNameMap.keys.toList(), reason).abortOnError()
+                rpDidToNameMap.forEach { relyingParty ->
+                    receiptRepository.createAndSaveReceiptsForVCs(
+                        relyingParty.key,
+                        relyingParty.value,
+                        ReceiptAction.Revocation,
+                        listOf(verifiableCredentialHolder.cardId)
+                    ).abortOnError()
+                }
+                Result.Success(Unit)
+            }
+        }
+    }
+
+    /**
+     * Revokes all verifiable presentations which revokes access for all relying parties to do a status check on the Verifiable Credential
+     *
+     * @param verifiableCredentialHolder The VC for which access to check status is revoked
+     * @param reason Reason for revocation
+     */
+    suspend fun revokeAllVerifiablePresentations(
+        verifiableCredentialHolder: VerifiableCredentialHolder,
+        reason: String = ""
+    ): Result<Unit> {
+        return withContext(Dispatchers.IO) {
+            runResultTry {
+                vchRepository.revokeVerifiablePresentation(verifiableCredentialHolder, emptyList(), reason).abortOnError()
+                //TODO: While adding advanced receipts improve on receipts for revoking all VPs
+                receiptRepository.createAndSaveReceiptsForVCs("", "", ReceiptAction.Revocation, listOf(verifiableCredentialHolder.cardId)).abortOnError()
                 Result.Success(Unit)
             }
         }
