+++ sdk/src/main/java/com/microsoft/did/sdk/credential/models/receipts/Receipt.kt
@@ -21,7 +21,7 @@ data class Receipt(
     val entityIdentifier: String,
 
     // date action occurred
-    val activityDate: Long,
+    val activityDate: Long = System.currentTimeMillis(),
 
     //Name of the verifier/issuer
     val entityName: String,
