+++ sdk/src/main/java/com/microsoft/did/sdk/credential/service/models/VcServiceActionRequest.kt
@@ -16,7 +16,7 @@ class ExchangeRequest(val verifiableCredential: VerifiableCredential, val pairwi
 class RevocationRequest(val verifiableCredential: VerifiableCredential, val owner: Identifier, val rpList: List<String>, val reason: String) :
     VcServiceActionRequest(verifiableCredential.contents.vc.revokeService?.id ?: "")
 
-typealias RevokedRPNameAndDid = Map<String, String>
+typealias RpDidToNameMap = Map<String, String>
 
 class StatusRequest(val verifiableCredential: VerifiableCredential, val owner: Identifier) :
     VcServiceActionRequest(verifiableCredential.contents.vc.credentialStatus?.id ?: "")
